+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 
 * nostr: add `TagKind::u` constructor ([Yuki Kishimoto])
 * nostr: derive `Copy` for `HttpMethod` ([Yuki Kishimoto])
+* nostr: add `nip98::verify_auth_header` ([Yuki Kishimoto])
 
 ### Fixed